IA-4748: code review
